refactor: 00029 refactor demo to lmax disruptor and 00055 persistence design doc (#41)

+++ server/src/test/java/com/hedera/block/server/util/TestUtils.java
@@ -17,18 +17,45 @@
 package com.hedera.block.server.util;
 
 import java.io.File;
+import java.nio.file.attribute.FileAttribute;
+import java.nio.file.attribute.PosixFilePermission;
+import java.nio.file.attribute.PosixFilePermissions;
+import java.util.Set;
 
 public final class TestUtils {
     private TestUtils() {}
 
-    public static void deleteDirectory(File directoryToBeDeleted) {
-        File[] allContents = directoryToBeDeleted.listFiles();
-        if (allContents != null) {
-            for (File file : allContents) {
-                deleteDirectory(file);
+    private static final String NO_PERMS = "---------";
+    private static final String NO_READ = "-wx-wx-wx";
+    private static final String NO_WRITE = "r-xr-xr-x";
+
+    public static boolean deleteDirectory(File directoryToBeDeleted) {
+
+        if (!directoryToBeDeleted.exists()) {
+            return true;
+        }
+
+        if (directoryToBeDeleted.isDirectory()) {
+            File[] allContents = directoryToBeDeleted.listFiles();
+            if (allContents != null) {
+                for (File file : allContents) {
+                    deleteDirectory(file);
+                }
             }
         }
 
-        directoryToBeDeleted.delete();
+        return directoryToBeDeleted.delete();
+    }
+
+    public static FileAttribute<Set<PosixFilePermission>> getNoPerms() {
+        return PosixFilePermissions.asFileAttribute(PosixFilePermissions.fromString(NO_PERMS));
+    }
+
+    public static FileAttribute<Set<PosixFilePermission>> getNoRead() {
+        return PosixFilePermissions.asFileAttribute(PosixFilePermissions.fromString(NO_READ));
+    }
+
+    public static FileAttribute<Set<PosixFilePermission>> getNoWrite() {
+        return PosixFilePermissions.asFileAttribute(PosixFilePermissions.fromString(NO_WRITE));
     }
 }
